feat(*): move slices to SM menu
Include slices in the SM menu so that they can be instantiated from there. Remove instances and requests from SP menu. Also, merge network services and instances in the same section.

Fixes #332
