+++ src/app/menu/menu.component.html
@@ -116,37 +116,41 @@
 				</button>
 			</div>
 
-			<button [ngClass]="{ active: section === 'sp-slices' }" (click)="setSection($event, 'sp-slices')"
-				i18n="@@slices">Slices</button>
-			<div class="section" *ngIf="section === 'sp-slices'">
+			<button [ngClass]="{ active: section === 'sp-slices' }" (click)="setSection($event, 'sp-slices')" i18n="@@networkSlices">Network
+				Slices</button>
+		</div>
+
+		<div class="options" *ngIf="menu === 'service-management'">
+			<button [ngClass]="{ active: section === 'sm-slices' }" (click)="setSection($event, 'sm-slices')" i18n="@@networkSlices">
+				Network Slices
+			</button>
+			<div class="section" *ngIf="section === 'sm-slices'">
 				<button [ngClass]="{ active: subsection === 'slices-templates' }" (click)="setSubsection($event, 'slices-templates')"
 					i18n="@@templates">
 					Templates
-				</button>
-				<button [ngClass]="{ active: subsection === 'slices-requests' }" (click)="setSubsection($event, 'slices-requests')"
-					i18n="@@requests">
-					Requests
 				</button>
 				<button [ngClass]="{ active: subsection === 'slices-instances' }" (click)="setSubsection($event, 'slices-instances')"
 					i18n="@@instances">
 					Instances
 				</button>
 			</div>
-		</div>
 
-		<div class="options" *ngIf="menu === 'service-management'">
 			<button [ngClass]="{ active: section === 'sm-network-services' }" (click)="setSection($event, 'sm-network-services')"
-				i18n="@@availableNS">
-				Available network services
+				i18n="@@nss">
+				Network Services
 			</button>
+			<div class="section" *ngIf="section === 'sm-network-services'">
+				<button [ngClass]="{ active: subsection === 'services' }" (click)="setSubsection($event, 'services')" i18n="@@services">
+					Services
+				</button>
+				<button [ngClass]="{ active: subsection === 'network-service-instances' }"
+					(click)="setSubsection($event, 'network-service-instances')" i18n="@@instances">
+					Instances
+				</button>
+			</div>
 
 			<button [ngClass]="{ active: section === 'sm-requests' }" (click)="setSection($event, 'sm-requests')"
 				i18n="@@requests">Requests</button>
-
-			<button [ngClass]="{ active: section === 'sm-network-service-instances' }"
-				(click)="setSection($event, 'sm-network-service-instances')" i18n="@@nsInstances">
-				Network service instances
-			</button>
 
 			<button [ngClass]="{ active: section === 'sm-licenses' }" (click)="setSection($event, 'sm-licenses')" i18n="@@licences">
 				Licences
