+++ src/app/service-management/slice-instance-create/slice-instance-create.component.scss
