+++ src/app/service-management/slice-instance-detail/slice-instance-detail.component.scss
