+++ src/app/service-management/slice-instance-list/slice-instance-list.component.scss
