+++ src/app/service-platform/slice-template-list/slice-template-list.component.ts
@@ -6,7 +6,7 @@ import { Subscription } from 'rxjs';
 import { ServicePlatformService } from '../service-platform.service';
 import { UtilsService } from '../../shared/services/common/utils.service';
 
-import { SliceInstanceCreateComponent } from '../slice-instance-create/slice-instance-create.component';
+import { SliceInstanceCreateComponent } from '../../service-management/slice-instance-create/slice-instance-create.component';
 
 @Component({
 	selector: 'app-slice-template-list',
